add option StrictHostKeyChecking=no for ssh so that the unknown host key can be saved automatic.
This is for solve the problem that when do repo sync, it will ask something like:
The authenticity of host 'xxx (x.x.x.x)' can't be established.
RSA key fingerprint is xxxxx.
Are you sure you want to continue connecting (yes/no)?

And then whatever you input yes or no, it can not continue.

After adding this option, it will add the unknow host key automatic and show a message below:
Warning: Permanently added 'xxx' (RSA) to the list of known hosts.

The repo sync can be continued.